+++ core/domain/search_services_test.py
@@ -14,16 +14,21 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+from core.domain import collection_services
 from core.domain import exp_services
 from core.domain import rating_services
 from core.domain import rights_manager
 from core.domain import search_services
 from core.domain import user_services
+from core.platform import models
 from core.tests import test_utils
+
+gae_search_services = models.Registry.import_search_services()
 
 class SearchServicesUnitTests(test_utils.GenericTestBase):
     """Test the search services module."""
     EXP_ID = 'An_exploration_id'
+    COLLECTION_ID = 'A_collection_id'
 
     def setUp(self):
         super(SearchServicesUnitTests, self).setUp()
@@ -102,3 +107,188 @@ class SearchServicesUnitTests(test_utils.GenericTestBase):
         exp_summary = exp_services.get_exploration_summary_by_id(self.EXP_ID)
         self.assertEqual(search_services.get_search_rank_from_exp_summary(
             exp_summary), 0)
+
+    def test_search_explorations(self):
+        expected_query_string = 'a query string'
+        expected_cursor = 'cursor'
+        expected_sort = 'title'
+        expected_limit = 30
+        expected_result_cursor = 'rcursor'
+        doc_ids = ['id1', 'id2']
+
+        def mock_search(query_string, index, cursor=None, limit=20, sort='',
+                        ids_only=False, retries=3):
+            self.assertEqual(query_string, expected_query_string)
+            self.assertEqual(index, search_services.SEARCH_INDEX_EXPLORATIONS)
+            self.assertEqual(cursor, expected_cursor)
+            self.assertEqual(limit, expected_limit)
+            self.assertEqual(sort, expected_sort)
+            self.assertEqual(ids_only, True)
+            self.assertEqual(retries, 3)
+
+            return doc_ids, expected_result_cursor
+
+        with self.swap(gae_search_services, 'search', mock_search):
+            result, cursor = search_services.search_explorations(
+                expected_query_string,
+                expected_limit,
+                sort=expected_sort,
+                cursor=expected_cursor,
+            )
+
+        self.assertEqual(cursor, expected_result_cursor)
+        self.assertEqual(result, doc_ids)
+
+    def test_patch_exploration_search_document(self):
+
+        def mock_get_doc(doc_id, index):
+            self.assertEqual(doc_id, self.EXP_ID)
+            self.assertEqual(index, search_services.SEARCH_INDEX_EXPLORATIONS)
+            return {'a': 'b', 'c': 'd'}
+
+        def mock_add_docs(docs, index):
+            self.assertEqual(index, search_services.SEARCH_INDEX_EXPLORATIONS)
+            self.assertEqual(docs, [{'a': 'b', 'c': 'e', 'f': 'g'}])
+
+        get_doc_swap = self.swap(
+            gae_search_services, 'get_document_from_index', mock_get_doc)
+
+        add_docs_counter = test_utils.CallCounter(mock_add_docs)
+        add_docs_swap = self.swap(
+            gae_search_services, 'add_documents_to_index', add_docs_counter)
+
+        with get_doc_swap, add_docs_swap:
+            patch = {'c': 'e', 'f': 'g'}
+            search_services.patch_exploration_search_document(
+                self.EXP_ID, patch)
+
+        self.assertEqual(add_docs_counter.times_called, 1)
+
+    def test_search_collections(self):
+        expected_query_string = 'a query string'
+        expected_cursor = 'cursor'
+        expected_sort = 'title'
+        expected_limit = 30
+        expected_result_cursor = 'rcursor'
+        doc_ids = ['id1', 'id2']
+
+        def mock_search(query_string, index, cursor=None, limit=20, sort='',
+                        ids_only=False, retries=3):
+            self.assertEqual(query_string, expected_query_string)
+            self.assertEqual(
+                index, collection_services.SEARCH_INDEX_COLLECTIONS)
+            self.assertEqual(cursor, expected_cursor)
+            self.assertEqual(limit, expected_limit)
+            self.assertEqual(sort, expected_sort)
+            self.assertEqual(ids_only, True)
+            self.assertEqual(retries, 3)
+
+            return doc_ids, expected_result_cursor
+
+        with self.swap(gae_search_services, 'search', mock_search):
+            result, cursor = search_services.search_collections(
+                expected_query_string,
+                expected_limit,
+                sort=expected_sort,
+                cursor=expected_cursor,
+            )
+
+        self.assertEqual(cursor, expected_result_cursor)
+        self.assertEqual(result, doc_ids)
+
+    def test_patch_collection_search_document(self):
+
+        def mock_get_doc(doc_id, index):
+            self.assertEqual(doc_id, self.COLLECTION_ID)
+            self.assertEqual(
+                index, search_services.SEARCH_INDEX_COLLECTIONS)
+            return {'a': 'b', 'c': 'd'}
+
+        def mock_add_docs(docs, index):
+            self.assertEqual(
+                index, search_services.SEARCH_INDEX_COLLECTIONS)
+            self.assertEqual(docs, [{'a': 'b', 'c': 'e', 'f': 'g'}])
+
+        get_doc_swap = self.swap(
+            gae_search_services, 'get_document_from_index', mock_get_doc)
+
+        add_docs_counter = test_utils.CallCounter(mock_add_docs)
+        add_docs_swap = self.swap(
+            gae_search_services, 'add_documents_to_index', add_docs_counter)
+
+        with get_doc_swap, add_docs_swap:
+            patch = {'c': 'e', 'f': 'g'}
+            search_services.patch_collection_search_document(
+                self.COLLECTION_ID, patch)
+
+        self.assertEqual(add_docs_counter.times_called, 1)
+
+    def test_update_private_collection_status_in_search(self):
+
+        def mock_delete_docs(ids, index):
+            self.assertEqual(ids, [self.COLLECTION_ID])
+            self.assertEqual(
+                index, search_services.SEARCH_INDEX_COLLECTIONS)
+
+        def mock_get_rights(unused_collection_id):
+            return rights_manager.ActivityRights(
+                self.COLLECTION_ID,
+                [self.owner_id], [self.editor_id], [self.viewer_id],
+                status=rights_manager.ACTIVITY_STATUS_PRIVATE
+            )
+
+        delete_docs_counter = test_utils.CallCounter(mock_delete_docs)
+
+        delete_docs_swap = self.swap(
+            gae_search_services, 'delete_documents_from_index',
+            delete_docs_counter)
+        get_rights_swap = self.swap(
+            rights_manager, 'get_collection_rights', mock_get_rights)
+
+        with get_rights_swap, delete_docs_swap:
+            search_services.update_collection_status_in_search(
+                self.COLLECTION_ID)
+
+        self.assertEqual(delete_docs_counter.times_called, 1)
+
+    def test_demo_collections_are_added_to_search_index(self):
+        results = search_services.search_collections('Welcome', 2)[0]
+        self.assertEqual(results, [])
+
+        collection_services.load_demo('0')
+        results = search_services.search_collections('Welcome', 2)[0]
+        self.assertEqual(results, ['0'])
+
+    def test_demo_explorations_are_added_to_search_index(self):
+        results, _ = search_services.search_explorations('Welcome', 2)
+        self.assertEqual(results, [])
+
+        exp_services.load_demo('0')
+        results, _ = search_services.search_explorations('Welcome', 2)
+        self.assertEqual(results, ['0'])
+
+    def test_update_private_exploration_status_in_search(self):
+
+        def mock_delete_docs(ids, index):
+            self.assertEqual(ids, [self.EXP_ID])
+            self.assertEqual(index, search_services.SEARCH_INDEX_EXPLORATIONS)
+
+        def mock_get_rights(unused_exp_id):
+            return rights_manager.ActivityRights(
+                self.EXP_ID,
+                [self.owner_id], [self.editor_id], [self.viewer_id],
+                status=rights_manager.ACTIVITY_STATUS_PRIVATE
+            )
+
+        delete_docs_counter = test_utils.CallCounter(mock_delete_docs)
+
+        delete_docs_swap = self.swap(
+            gae_search_services, 'delete_documents_from_index',
+            delete_docs_counter)
+        get_rights_swap = self.swap(
+            rights_manager, 'get_exploration_rights', mock_get_rights)
+
+        with get_rights_swap, delete_docs_swap:
+            search_services.update_exploration_status_in_search(self.EXP_ID)
+
+        self.assertEqual(delete_docs_counter.times_called, 1)
